Development

+++ src/test/java/io/github/mat3e/HelloServiceTest.java
@@ -69,7 +69,7 @@ public class HelloServiceTest {
         //given
         var mockRepository = new LangRepository(){
             @Override
-            Optional<Lang> findById(final Long id) {
+            Optional<Lang> findById(final Integer id) {
                 return Optional.empty();
             }
         };
@@ -85,7 +85,7 @@ public class HelloServiceTest {
     private LangRepository fallbackLangIdRepository() {
         return new LangRepository() {
             @Override
-            Optional<Lang> findById(final Long id) {
+            Optional<Lang> findById(final Integer id) {
                 if (id.equals(HelloService.FALLBACK_LANG.getId())) {
                     return Optional.of(new Lang(null, FALLBACK_ID_WELCOME, null));
                 }
@@ -97,7 +97,7 @@ public class HelloServiceTest {
     private LangRepository alwaysReturnHelloRepository() {
         return new LangRepository() {
             @Override
-            Optional<Lang> findById(final Long id) {
+            Optional<Lang> findById(final Integer id) {
                 return Optional.of(new Lang(null, WELCOME, null));
             }
         };
